Add rake tasks to build and publish our site
To move away from github hosting, we need tasks to build our stuff and
publish it.

This doesn't yet have the necessary safety checks to keep you from doing
things like "pushing a random branch to production" but, as a
compensating control, only staging is grabbing things from the webroot
we specified at this point.

To use:

Run
`rake build deploy` to deploy to the staging env, assuming your key is in
the right places.

Run
`rake build deploy_prod` to deploy to the production env, assuming your
key is in the right places and we've edited nginx to refer to the hosted
root instead of the proxy.