Switch over to relative links for CSP header compliance
A bunch of our stuff was being hard-coded as https://baseurl/assets or,
more confusingly, https://lucyparsonslabs.com/images/, which made trying
to make CSP headers that didn't break the site and weren't restrictive
annoyingly difficult.

Since the site is accessible as both www.lucyparsonslabs.com and
lucyparsonslabs.com, we might as well serve up relative links to the
content in question. _config.yml has the meat of it, but some changes
were required in the header and footer_scripts to serve the stuff out of
where we wanted it to come from.

We can now live without fear of baseurl.

Also fixed the Black Rose blogpost's thumbnail, since it was 404ing.